requirements.txt
aiogram>=3.0.0,<4.0.0
python-dotenv>=1.0.0
pytz>=2020.0
cryptography>=39.0.0
pydantic-settings
